backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djoser>=2.2.0
Pillow>=10.0.0
python-multipart>=0.0.6
pytesseract>=0.3.10
PyMuPDF>=1.23.0
pdf2image>=1.16.0
